Add effort values 0 and 0.5 with fixed Fibonacci scale
Widens PBI/Story/Bug effort estimation from positive-integers-only to the
fixed scale [0, 0.5, 1, 2, 3, 5, 8, 13, 21]. Zero and half-point stories
were previously rejected by a gt=0 validator and an integer DB column.

- Alembic migration: pbis.effort INTEGER → REAL (batch mode, data-safe)
- Backend schemas: field_validator enforces the allowed set; PBIResponse,
  FeatureResponse, SprintResponse, SwimlineResponse, PIResponse all widened
  to float; PBIResponse.item_type tightened to Literal["story","bug"]
- CSV import: locale-aware before-validator accepts "0,5" comma decimal
- Effort service: int() casts changed to float() to preserve 0.5 sums
- MCP tools: effort params updated to float with corrected descriptions
- Frontend: number input replaced with button-group (½ for 0.5); fixed
  falsy-zero bug (effort || null → effort ?? null); csvParser rewritten
  to validate against the same allowed set; SprintCreate defined locally
  in index.ts (no longer emitted by the OpenAPI spec)
- Tests: fixed 4 existing tests using disallowed values (4, 6, 7);
  added boundary tests for 0, 0.5, invalid values, and locale CSV input

+++ backend/app/models/pbi.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime
 from uuid import uuid4
 
-from sqlalchemy import ForeignKey, Index, Integer, Text, UniqueConstraint, func
+from sqlalchemy import Float, ForeignKey, Index, Integer, Text, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -18,7 +18,7 @@ class PBI(Base):
     parent_feature_system_id: Mapped[str] = mapped_column(Text, ForeignKey("features.system_id", ondelete="CASCADE"), nullable=False)
     title: Mapped[str] = mapped_column(Text, nullable=False)
     description: Mapped[str | None] = mapped_column(Text)
-    effort: Mapped[int | None] = mapped_column(Integer)
+    effort: Mapped[float | None] = mapped_column(Float)
     item_type: Mapped[str] = mapped_column(Text, nullable=False, default="story")
     location: Mapped[str] = mapped_column(Text, nullable=False, default="backlog")
     pi_id: Mapped[str | None] = mapped_column(Text, ForeignKey("pis.system_id"))
